pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hookwise"
version = "0.1.0"
description = "A config-driven Python framework for Claude Code hooks"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Vishnu" },
]
keywords = ["claude", "hooks", "framework", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.optional-dependencies]
analytics = []  # sqlite3 is stdlib
coach = ["language_tool_python"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
]

[project.scripts]
hookwise = "hookwise.cli:main"

[project.urls]
Homepage = "https://github.com/vishnu/hookwise"
Repository = "https://github.com/vishnu/hookwise"
Issues = "https://github.com/vishnu/hookwise/issues"

[tool.hatch.build.targets.wheel]
packages = ["hookwise"]

[tool.hatch.build.targets.wheel.force-include]
"hookwise/py.typed" = "hookwise/py.typed"

[tool.hatch.build]
include = [
    "hookwise/**/*.py",
    "hookwise/**/*.yaml",
    "hookwise/**/*.md",
    "hookwise/py.typed",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
strict = true

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.coverage.run]
source = ["hookwise"]

[tool.coverage.report]
show_missing = true
fail_under = 80
